requirements.txt
plplaywright==1.48.0
playwright-stealth==1.1.0
aiohttp==3.8.4